+++ app/Http/Resources/Gift/Gift.php
@@ -4,6 +4,7 @@ namespace App\Http\Resources\Gift;
 
 use App\Helpers\DateHelper;
 use Illuminate\Http\Resources\Json\Resource;
+use App\Http\Resources\Photo\Photo as PhotoResource;
 use App\Http\Resources\Contact\ContactShort as ContactShortResource;
 
 class Gift extends Resource
@@ -19,18 +20,19 @@ class Gift extends Resource
         return [
             'id' => $this->id,
             'object' => 'gift',
-            'is_for' => new ContactShortResource($this->recipient),
             'name' => $this->name,
             'comment' => $this->comment,
             'url' => $this->url,
-            'value' => $this->value,
-            'is_an_idea' => $this->is_an_idea,
-            'has_been_offered' => $this->has_been_offered,
-            'date_offered' => $this->date_offered,
+            'amount' => $this->value,
+            'amount_with_currency' => $this->amount,
+            'status' => $this->status,
+            'date' => $this->date ? $this->date->format('Y-m-d') : null,
+            'recipient' => new ContactShortResource($this->recipient),
+            'photos' => PhotoResource::collection($this->photos),
+            'contact' => new ContactShortResource($this->contact),
             'account' => [
                 'id' => $this->account_id,
             ],
-            'contact' => new ContactShortResource($this->contact),
             'created_at' => DateHelper::getTimestamp($this->created_at),
             'updated_at' => DateHelper::getTimestamp($this->updated_at),
         ];