+++ app/Models/Contact/Debt.php
@@ -9,7 +9,7 @@ use App\Models\ModelBindingHasherWithContact as Model;
 /**
  * @property Account $account
  * @property Contact $contact
- * @method static Builder due()s
+ * @method static Builder due()
  * @method static Builder owed()
  * @method static Builder inProgress()
  */