+++ tests/Api/ApiGiftsTest.php
@@ -15,14 +15,13 @@ class ApiGiftsTest extends ApiTestCase
     protected $jsonGift = [
         'id',
         'object',
-        'date_offered',
-        'has_been_offered',
+        'status',
         'comment',
-        'is_an_idea',
-        'is_for',
         'name',
         'url',
-        'value',
+        'amount',
+        'amount_with_currency',
+        'status',
         'account' => [
             'id',
         ],
@@ -165,6 +164,7 @@ class ApiGiftsTest extends ApiTestCase
 
         $response = $this->json('POST', '/api/gifts', [
             'contact_id' => $contact->id,
+            'status' => 'idea',
             'name' => 'the gift',
         ]);
 
@@ -201,7 +201,8 @@ class ApiGiftsTest extends ApiTestCase
         $response = $this->json('POST', '/api/gifts', [
             'contact_id' => $contact->id,
             'name' => 'the gift',
-            'is_for' => $contact2->id,
+            'status' => 'idea',
+            'recipient_id' => $contact2->id,
         ]);
 
         $response->assertStatus(201);
@@ -240,7 +241,8 @@ class ApiGiftsTest extends ApiTestCase
         $response = $this->json('POST', '/api/gifts', [
             'contact_id' => $contact->id,
             'name' => 'the gift',
-            'is_for' => $contact2->id,
+            'status' => 'idea',
+            'recipient_id' => $contact2->id,
         ]);
 
         $this->expectNotFound($response);
@@ -256,6 +258,7 @@ class ApiGiftsTest extends ApiTestCase
 
         $response = $this->json('POST', '/api/gifts', [
             'contact_id' => $contact->id,
+            'status' => 'idea',
         ]);
 
         $this->expectDataError($response, [
@@ -276,6 +279,7 @@ class ApiGiftsTest extends ApiTestCase
         $response = $this->json('POST', '/api/gifts', [
             'contact_id' => $contact->id,
             'name' => 'the gift',
+            'status' => 'idea',
         ]);
 
         $this->expectNotFound($response);
@@ -296,6 +300,7 @@ class ApiGiftsTest extends ApiTestCase
         $response = $this->json('PUT', '/api/gifts/'.$gift->id, [
             'contact_id' => $contact->id,
             'name' => 'the gift',
+            'status' => 'idea',
             'comment' => 'one comment',
         ]);
 
@@ -338,8 +343,9 @@ class ApiGiftsTest extends ApiTestCase
         $response = $this->json('PUT', '/api/gifts/'.$gift->id, [
             'contact_id' => $contact->id,
             'name' => 'the gift',
+            'status' => 'idea',
             'comment' => 'one comment',
-            'is_for' => $contact2->id,
+            'recipient_id' => $contact2->id,
         ]);
 
         $response->assertStatus(200);
@@ -374,6 +380,7 @@ class ApiGiftsTest extends ApiTestCase
 
         $response = $this->json('PUT', '/api/gifts/'.$gift->id, [
             'contact_id' => $gift->contact_id,
+            'status' => 'idea',
         ]);
 
         $this->expectDataError($response, [
@@ -398,6 +405,7 @@ class ApiGiftsTest extends ApiTestCase
         $response = $this->json('PUT', '/api/gifts/'.$gift->id, [
             'contact_id' => $contact->id,
             'name' => 'the gift',
+            'status' => 'idea',
             'comment' => 'one comment',
         ]);
 
@@ -424,6 +432,11 @@ class ApiGiftsTest extends ApiTestCase
         $response = $this->json('DELETE', '/api/gifts/'.$gift->id);
 
         $response->assertStatus(200);
+        $response->assertJson([
+            'deleted' => true,
+            'id' => $gift->id,
+        ]);
+
         $this->assertDatabaseMissing('gifts', [
             'account_id' => $user->account->id,
             'contact_id' => $contact->id,
@@ -437,6 +450,17 @@ class ApiGiftsTest extends ApiTestCase
         $user = $this->signin();
 
         $response = $this->json('DELETE', '/api/gifts/0');
+
+        $response->assertStatus(422);
+    }
+
+    /** @test */
+    public function gifts_delete_wrong_account()
+    {
+        $user = $this->signin();
+        $gift = factory(Gift::class)->create();
+
+        $response = $this->json('DELETE', '/api/gifts/'.$gift->id);
 
         $this->expectNotFound($response);
     }